Back to commiting generated files to repository, but in separate folder. (#9)
The idea of switching to artifacts was to separate the automatic output from the source code and only track changes made by users. However, I did not consider that the script efficiently only generates the output of changed files, such that we end up with artifacts that only contain recent changes, but not a snapshort of the current state of the db. The compromise now is to commit the generated files, but put them into a "generated" subdir, so it remains clear which files are source and which are automatically generated.